+++ Packages/com.fang.framework/Runtime/Core/ILifecycle.cs
@@ -1,0 +1,19 @@
+namespace Fang.Framework
+{
+    public interface ILifecycle
+    {
+        void OnInit();
+
+        void OnDispose();
+    }
+
+    public interface ITickable
+    {
+        void OnTick(float deltaTime);
+    }
+
+    public interface IFixedTickable
+    {
+        void OnFixedTick(float deltaTime);
+    }
+}