+++ Packages/com.fang.framework/Runtime/Service/Service.cs
@@ -1,6 +1,0 @@
-namespace Fang.Framework
-{
-    public abstract class Service
-    {
-    }
-}