+++ Packages/com.fang.framework/Tests/Runtime/ScopeTestDoubles.cs
@@ -1,58 +1,7 @@
-using System;
 using System.Collections.Generic;
 
 namespace Fang.Framework.Tests
 {
-    internal interface IProbeService
-    {
-    }
-
-    internal sealed class ProbeService : IProbeService
-    {
-    }
-
-    internal sealed class ProbeServiceOverride : IProbeService
-    {
-    }
-
-    internal sealed class ProbeFactoryResult : IProbeService
-    {
-    }
-
-    internal sealed class UnregisteredService
-    {
-    }
-
-    internal sealed class DependentService
-    {
-        public DependentService(IProbeService probe)
-        {
-            Probe = probe;
-        }
-
-        public IProbeService Probe { get; }
-    }
-
-    internal sealed class CircularServiceA
-    {
-        public CircularServiceA(CircularServiceB other)
-        {
-            Other = other;
-        }
-
-        public CircularServiceB Other { get; }
-    }
-
-    internal sealed class CircularServiceB
-    {
-        public CircularServiceB(CircularServiceA other)
-        {
-            Other = other;
-        }
-
-        public CircularServiceA Other { get; }
-    }
-
     internal static class LifecycleLog
     {
         public static readonly List<string> Entries = new List<string>();
@@ -68,72 +17,143 @@ namespace Fang.Framework.Tests
         }
     }
 
-    internal sealed class LifecycleProbeA : IInjectable, IInitializable, IDisposable
+    internal class ProbeScope : Scope
+    {
+        public TService Add<TService>() where TService : Service, new()
+        {
+            return AddService<TService>();
+        }
+
+        public void Remove<TService>() where TService : Service
+        {
+            RemoveService<TService>();
+        }
+
+        public TChild AddChild<TChild>() where TChild : Scope, new()
+        {
+            return CreateChildScope<TChild>();
+        }
+    }
+
+    internal sealed class ProbeChildScope : ProbeScope
+    {
+    }
+
+    internal sealed class ProbeGrandChildScope : ProbeScope
+    {
+    }
+
+    internal sealed class AudioService : Service
+    {
+        public Scope InjectedScope => Scope;
+    }
+
+    internal sealed class AudioServiceOverride : Service
+    {
+    }
+
+    internal sealed class LifecycleProbeA : Service
     {
         public LifecycleProbeA()
         {
             LifecycleLog.Add("create:A");
         }
 
-        public void Inject(Scope scope)
+        public override void OnInit()
         {
-            LifecycleLog.Add("inject:A");
+            LifecycleLog.Add("init:A");
         }
 
-        public void Initialize()
-        {
-            LifecycleLog.Add("initialize:A");
-        }
-
-        public void Dispose()
+        public override void OnDispose()
         {
             LifecycleLog.Add("dispose:A");
         }
     }
 
-    internal sealed class LifecycleProbeB : IInjectable, IInitializable, IDisposable
+    internal sealed class LifecycleProbeB : Service
     {
         public LifecycleProbeB()
         {
             LifecycleLog.Add("create:B");
         }
 
-        public void Inject(Scope scope)
+        public override void OnInit()
         {
-            LifecycleLog.Add("inject:B");
+            LifecycleLog.Add("init:B");
         }
 
-        public void Initialize()
-        {
-            LifecycleLog.Add("initialize:B");
-        }
-
-        public void Dispose()
+        public override void OnDispose()
         {
             LifecycleLog.Add("dispose:B");
         }
     }
 
-    internal sealed class LifecycleProbeC : IInjectable, IInitializable, IDisposable
+    internal sealed class LifecycleProbeC : Service
     {
         public LifecycleProbeC()
         {
             LifecycleLog.Add("create:C");
         }
 
-        public void Inject(Scope scope)
+        public override void OnInit()
         {
-            LifecycleLog.Add("inject:C");
+            LifecycleLog.Add("init:C");
         }
 
-        public void Initialize()
-        {
-            LifecycleLog.Add("initialize:C");
-        }
-
-        public void Dispose()
+        public override void OnDispose()
         {
             LifecycleLog.Add("dispose:C");
+        }
+    }
+
+    internal sealed class TickProbeService : Service, ITickable, IFixedTickable
+    {
+        public void OnTick(float deltaTime)
+        {
+            LifecycleLog.Add("tick:first");
+        }
+
+        public void OnFixedTick(float deltaTime)
+        {
+            LifecycleLog.Add("fixedTick:first");
+        }
+    }
+
+    internal sealed class SecondaryTickProbeService : Service, ITickable, IFixedTickable
+    {
+        public void OnTick(float deltaTime)
+        {
+            LifecycleLog.Add("tick:second");
+        }
+
+        public void OnFixedTick(float deltaTime)
+        {
+            LifecycleLog.Add("fixedTick:second");
+        }
+    }
+
+    internal sealed class FixedOnlyProbeService : Service, IFixedTickable
+    {
+        public void OnFixedTick(float deltaTime)
+        {
+            LifecycleLog.Add("fixedTick:only");
+        }
+    }
+
+    internal sealed class DeltaProbeService : Service, ITickable, IFixedTickable
+    {
+        public float LastTick { get; private set; }
+
+        public float LastFixedTick { get; private set; }
+
+        public void OnTick(float deltaTime)
+        {
+            LastTick = deltaTime;
+        }
+
+        public void OnFixedTick(float deltaTime)
+        {
+            LastFixedTick = deltaTime;
         }
     }
 }